fix length jwt

+++ frontend/src/page/Zalo.jsx
@@ -26,28 +26,50 @@ export default function Zalo() {
     setChat(true);
   };
 
+  // useEffect(() => {
+  //   const fetch = async () => {
+  //     const response = await axios.post(
+  //       "http://127.0.0.1:8080/auth/token",
+  //       {},
+  //       { withCredentials: true }
+  //     );
+  //     console.log(response);
+  //     if (response.status === 200) {
+  //       setUserData(response.data);
+  //       setChat(true);
+  //       return;
+  //     }
+  //     setChat(false);
+  //     setIsLoadding(true);
+  //   };
+  //   fetch();
+  // }, []);
+
   useEffect(() => {
-    const fetch = async () => {
-      const response = await axios.post(
-        "http://127.0.0.1:8080/auth/token",
-        {},
-        { withCredentials: true }
-      );
-      console.log(response);
+    const fetchLoginToken = async () => {
+      const url = "http://localhost:8080/auth/token";
+      const response = await axios.post(url, {}, { withCredentials: true });
+
       if (response.status === 200) {
+        console.log(response);
         setUserData(response.data);
         setChat(true);
-        return;
+      } else {
+        setChat(false);
       }
-      setChat(false);
-      setIsLoadding(true);
+      setIsLoadding(false);
     };
-    fetch();
+    fetchLoginToken();
   }, []);
 
-  const handleLogout = () => {
-    localStorage.clear();
-    setChat(false);
+  const handleLogout = async () => {
+    const url = "http://localhost:8080/auth/logout";
+    const response = await axios.post(url, {}, { withCredentials: true });
+
+    if (response.status === 200) {
+      localStorage.clear();
+      setChat(false);
+    }
   };
 
   return (
